New kca-blueprints-ready utility
Given my blueprint info and picture book info, this tells me which ships
are ready for a modification upgrade, and which are still waiting for
more blueprints to drop.

Even without knowing about ships I haven't seen yet, it's pretty useful.

Once it has a complete ship-list from the Wiki, it should also be useful
for reporting expiring and unneeded blueprints for end-of-the-month too.

And it's shaking out more APIs to consider moving into the library.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>