Use lazy_static_include to read test data fixtures
This removes the dependency on having the CARGO_MANIFEST_DIR env-var
available at test-time, and means that `cargo wasi test --release` works
without configuration on both wasmtime and wasmedge.

However, I suspect this embeds the full path to the file into the
executable, which breaks the existing workaround I was using for getting
`cargo wasi test` to work (mapping the PWD to root), but I can't tell
for sure due to lack of useful output when failing due to the WASI
platform being terminate-on-panic. The stack trace suggests that a
different test is failing, which would make sense if they're being run
in parallel, but makes it hard to diagnose.

Anyway, it's an acceptible loss: This is only best-effort testing until
I can migrate to cargo-component, wasip2, and perhaps a better test
integration.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>