Give new CLI parser tests unique function names
rust-analyzer Test Explorer feature seems to mishandle the case of
multiple tests with the same function name, and even though it shows the
tests in the hierarchy, their results are all assigned to the first one
in the list.

Either that, or I was doing something wrong. Need to put together a
minimal reproduction and report it upstream.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>