Introduce Ship abstraction layer
This is intended to be the primary API for this library, combining all
data sources into one reasonably useable API.

I don't know if I'll keep the direct access to underlying objects or put
some real getters around them.

Once I have converted all the existing utilities to use this API, I can
look at simplifying the underlying data holders, the existing setup that
prevents modification was wrong-headed at that level.

(I'm repeating that mistake at the higher level, but it'll be easier to
clean up if and when that turns out to be wrong-headed too.)

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>