Move remaining CLI tools to use Ships abstraction
This lets us un-public the individual importer functions, ensuring there
are no users that cannot use the Ships abstraction, and hence allowing
further internal refactoring/simplification of the underlying
unabstracted systems.

Long-term, there's no strong need for the importers to be private, but I
am leaning towards keeping things private as much as possible while the
whole API is still in-flux, to ensure we get a useful abstraction.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>