Report cost 0 for ships with no possible upgrade
最上改二 and 最上改二特(改三 level) exist, but since they are not yet
constructable, there's no 重巡洋艦 with such levels, and so there's no
data for that case.

The current output is intended to allow checking for incorrectly-marked
ships, i.e. when a further upgrade exists in-game but the tool doesn't
have the data. So falling back to 0 should be clear-enough in the output.

Once the data is "reliable", i.e. full ship list sourced from the Wiki,
then this sort of workaround can be revisited, as will the whole logic
of this tool.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>