Rationalise integration test data into one folder
This will make it easier to update in future, and removes a bunch of
currently-unused data files. I can always bring them back if they are
missed, but by-and-large I'm not invested in supporting old data,
particularly if formats change or invariants vary.

The intent is to use this library with live data, after all.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>