Fix logic for ships found via book or blueprint
Rather than skip any ship where we have had a book or blueprint match,
only skip ships where using the book or blueprint entry would create an
entry that already exists.

This should correctly handle cases where the player has gained ships or
variants not known to the kekkon list but which are visible in the book
or blueprint list. Previously, such ships would not have an entry in the
Ships list.

This case would only be triggered by events which can reward modified
ships outside the usual modified-ship drop requirements. However,
someone who bought a 改二 ship card offline without owning the base ship,
and then collecting a blueprint from a drop of the same ship would be in
this position.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>