Simplify fallback ships with `HashMap::entry`
I'm leaning towards actually flipping this function around and walking
each list, using `entry` to find a Ship or Ships to attach a given data
point to. It'd be simpler and more-robust, and probably not meaningfully
inefficient.

However, borrowing rules may prevent this if I want to move a value in
both `and_modify` and `or_insert_with_key`, as the borrow-checker may
not know that only one of those lambdas is called.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>