Implemented CharacterList importer
This is the final piece of per-ship data from the API. We're not using
it yet, but the plan is to use star counts to enhance team building, and
it turns out the display sort id might be useful to verify blueprint
matches, since they include the ship number of the base ship.

It also points to refactoring being needed in the Ships construction
code as adding new data sources is needlessly fiddly and complex.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>