Refactor Ships::new to be less painful to extend
Now we walk each set of input lists, and modify Ship values in-place if
they already exist, rather than trying to build the perfect Ship value
the first time we see a ship.

This is both easier to extend with new data sets, and should be less
cloney in the long run.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>