Refactor Ships::new to avoid cloning data objects
`and_modify` forces us to clone the thing we want to add to the object
since we also need to use it in `or_insert_with_key`, and there's no
obvious way to tell Rust that only one or the other closure will be
called.

`or_insert_with_key` returns a mutable reference to the entry value, so
we don't need `and_modify` anyway, we can just modify the object after
possibly inserting it.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>