Rename DecisiveBattle to CapeEngano
An エンガノ岬決戦mode outfit has been released for 瑞鶴改二甲, and while
looking into this, I discovered that the existing DecisiveBattle outfit
for 島風 is also called エンガノ岬決戦mode in the web version of the game.
There's also a bunch more ship-girls with エンガノ岬決戦mode outfits.

It seems "DecisiveBattle" (決戦mode) is a catch-call term for both
CapeEngano and SurigaoStrait modes in Twitter hash-tags and when
figurines are released, which confused web-searches a bit.

These are apparently the only two limited-time event areas in the web
version to get custom outfits. the other pages are tied to annual
repeating events. It seems the OriginalIllustration concept is
Arcade-only.

I don't know if the arcade version never held a CapeEngano event, or if
it happened but didn't include the new outfits for some reason. The
SurigaoStrait event was marked as the first part of the Battle of Leyte
Gulf so we may get a CapeEngano event later.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>