Parse ship-lists from kancolle-a wikwiki
Abusing a CSV parser that had serde support turned out to be much easier
than trying to implement a serde deserialiser.

As it happens, I have in my collection all of the ships that aren't in
the kekkon list (or at least an unmodified version that shares a book
entry) so this data doesn't do much for me, but might help future users,
as assuming the wiki is correct, there's exactly four characters which
are missing from the kekkon list.

I have not implemented CLI support for custom wiki lists, 'cause I don't
think anyone will be interested in using that. There's not a wild amount
of data gained from it anyway; particularly, ship upgrade ordering is
not available, so we will still need to build this from reliable
character ordering data and multi-row book entries, then name-based
guessing for unowned characters, i.e. pulling the logic from the
kca-blueprints-ready tool back into the Ships data structure.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>