Restructure as a Cargo Workspace
This change is mainly to separate the cli tools out so they can pull
dependencies that we don't want in the main library, for both size and
constrained platform, e.g., WASI, reasons.

I hope to add a GUI tool of some kind, and this is the first step in
preparing for that.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>