Reinstate compilation/testing under WASM targets
This compiles with both `--target wasm32-wasip1` and `cargo component`,
and `cargo test --target wasm32-wasip1 --release` runs the tests if you
have a runner configured, e.g.,
> $Env:CARGO_TARGET_WASM32_WASIP1_RUNNER="wasmtime"

Non-release builds don't work because I can't work out the right magic
to get `lazy_static_include_bytes!` working with the wasmtime runtime,
and there's basically zero debug output.

The CLI tools don't work because we don't currently have WASM support
for setting cookies (for --jsessionid) nor for a TTY (for password
prompting with rpassword), and even if I hardcode a password, reqwest
relies on being able to farm out its work to the JS `fetch` API, which
relies on a wasm_bindgen, which I don't currently have.

I'd prefer to disable the kancolle-a-cli-tools crate entirely under the
WASM targets, but I don't see a way to do that in the workspace config.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>