Cleanly-separate WASM-specific code and test in CI
I broke the imports on wasm while implementing thiserror, and didn't
notice: So it's time to add tests to CI!

The tests are now running on Windows with wasmedge 0.14.1 in non-release
mode. It turns out the failures seen previously were because wasmedge
before 0.14.1 was not running its test suite on Windows, and always
failed in things tried to read from EOF'd files, e.g.,
File::read_to_end or File::read_to_string.

CI is showing failures for the same case on Linux though, so CI will be
run in release mode for wasm32-wasip1.

This broke lazy_static_include_bytes! in non-release builds, as it uses
fs::read, which is implemented using io::default_read_to_end (not
File::read, despite the name).

I also took the opportunity to move the cfg-split code into a sub-module
so we only have one attribute checking the cfg option, for maintenance
and readability. Not sure if it's Rust-idiomatic, but I have enough Go
programming time that it looks fine to me.

Credit to https://chrismorgan.info/blog/rust-cfg_attr-mod-path/ for the
idea.

I also changed the split cfg option to be target_arch, since that's what
is checked by rewqest for its cookie-access API. The wasi-http spec
doesn't support cookie manipulation, so long-term we may want to drop
JSESSION input from the API and simply enable the cookie store on
non-wasm32 platforms as a one-liner, and rely on host default behaviour
on wasm32 platforms.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>