Restructure Ships to be a list of ShipMods
This rationalises the behaviour of the anything dealing with blueprints
or modification, and anything that wanted to walk all characters can use
the new shipmods_iter function to retain existing behaviour but with all
modifications of the same character grouped and in-order.

kca-blueprints-ready needed to be rewritten, but that was the main
motiviation for this, along with some kind of future tool that can tell
you quickly if a given gacha pull is new to your collection or not.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>