Add --character-only to kca-renovation-priority
This changes the meaning of "missing" to only consider whether the
character is in the library; even if you are missing the normal card for
that ShipMod, it won't be considered missing.

This lets me focus on unlocking new characters before backfilling book
entries, since blueprints can be used for the latter more-easily.

This logic will probably be inverted when I add support for time-limited
card pages, and in the meantime, kca-missing-card can be used to fill
the gap.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>