Simplify BookShip usage by splitting them
ShipMods which represent two book entries will now have a BookShip which
has been modified to contain only the relevant card images, status
images, and married status and images.

This removes the need to callers to handle the special case of 雪風 and
雪風改, and will also allow removing ShipMod.book_secondrow as it should
now always be false.

The downside to this is we have to deal with cases that cannot appear in
the original data, i.e. a card page with variation_num_in_page of 0. In
the future, we may be able to instead remove such card pages after
splitting in order to reinstatiate the expected behaviour.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>