Don't skip shipmod pairs when the left is special
This fixes a bug where kca-renovation-priority was not listing 阿武隈改
as READY for renovation to 阿武隈改二 because I don't have the normal base
card for 阿武隈改.

I have holo or special versions of both 阿武隈改 and 阿武隈改二, so in
character-only mode neither appears as missing.

As far as I know this is the only case of this in my data, but it
happens I've been tracking 阿武隈 so I knew what I was looking for.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>