Show STARS NEEDED for last mod even if drop needed
This ensures that we don't hide STARS NEEDED status when we are missing
the Normal page's first card.

The logic still needs a bit of work, as we do not currently distinguish
in the upgrade statuses if the upgrade is because the upper character is
missing or just the upper character's normal card.

This would probably be better with a GUI...

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>