Update for VERSION E REVISION 2 (2024/10/31)
I meant to do an API dump yesterday, and then one this morning, so that
I could have a clean commit that was only the API changes and any new
ships (there's two new book entries, which matches the announcement) but
I forgot.

So instead, the API dump also includes the last few weeks of progress,
and I had fixed a few holes in my collection over that time.

This version adds a field to the Character list, `slotExtension` but so
far it's `false` in all my data; it may be related to the two new ships,
as they are both American and hence may have distinct features.

This invalidates the existing character list test fixture, sadly, but we
only had one fixed-time capture to invalidate so it's not too bad.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>