Move ship_blueprint_costs into the library
I kept the function as-is, but wrapped it up in an associated function
of Ship to hide the need to hunt through the various data structures to
confirm the ship name and type for the blueprints if called on a ship
that has no blueprints.

This changes the behaviour of kca-blueprints-ready because we no longer
report ships that _might_ have a next stage. This is because our static
data now includes a Wiki dump so it's only going to be wrong during the
period between a new remodel level for a ship being introduced and when
it is added to the Wiki.

New ships being introduced and being immediately blueprint-constructible
is much rarer than day-to-day "What ships am I working on" inquiries,
but it does happen, e.g., 時雨改三 arrived this way (as a possible drop
during the Swimsuit festival). For contrast, 最上改二 was not
blueprintable when introduced, as she was a special-stage reward in a
limited-time mission; she became blueprint-constructible a few months
later.

It just occurred to me that I could store the current TCBook count
as a kind of data-versioning thing, since all players see the same count
of "not-implemented" cards. Same for static Wiki and kekkonkakkokari...

I also took the opportunity to add unit tests for ship_blueprint_name,
as I had previously overlooked doing so when moving it into the library.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>