Add test to show changes in TCBook/Wiki divergence
Basically if this changes, either the Wiki data has been updated, and
the test needs to be updated, or new ships have been released and the
wiki and/or test need to be updated.

Perhaps this should be an integration test, but we don't currently
expose unowned ships in the TCBook data's public exposure (ships module)
in order to make the main use-cases of "Is this in the TCBook" tests
more rational.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>