Data update: 雪風改二 (#456) added, events updated
The previous event has ended, and an Aime Campaign has started as well.

Interestingly, the event name doesn't appear to be in the data. I guess
it's hard-coded in the website. They also keep the event info live for
a little while after the event. So now I have the final state of this
event in my commit history.

Sadly, I missed out on Gambier Bay改, and finished with two blueprints
for her plus a second basic Iowa. >_<

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>