WIP HACK: Sorting for ships being output
I needed something to get sorted output when looking at a list of 102
blueprints across probably 80 ships, so I can use it as a one-pass
picklist.

I need to think about this more, both in terms of implementation of sort
order, and also how to expose it ergonomically and efficiently.

Maybe ShipBuilder should be able to be given the sort order function? We
could have functions for different sorts: Book order of base, book order
of highest (owned/known) renovation,
disp_sort_no (Char only) of base,
ship_sort_no (BP and Char, is this just shipType?), etc.

(Those last two aren't great, per below research)

So probably want code that pulls the Wiki's page that I use for sorting,
to populate that sort order (WIKI_SORT_ORDER here).

Using something like https://crates.io/crates/indexmap, then sorting can
be done after building the collection, and doesn't move the objects when
sorting, but is effectively keeping a sort around. So Ships can be
default-sorted, and then if callers want a different order, using
sorted_unstable_by to iterate should be relatively cheap.

Probably makes sense to allow providing a default sort comparison
function to ShipsBuilder anyway, while its mutable.

And maybe time to consider code that merges BP data and tries to build a
single view of the BP-only data? I have probably 80% blueprint coverage,
and it's possible that disp_sort_no can be reconstructed from blueprints
perhaps.

dispSortNo is, per "Breakdown of dispSortNo"
bookNo (not in blueprint list)
shipClassId (we get a shipClass string in Character)
- string changes during remodeling, e.g., U-511, but shipClassId doesn't
shipClassIndex (Also directly in character, but not always populated)
remodelLv

So if we want to sort by this, rather than book_id, then we just take
the middle and it's sortable. Not sure if it's a good sort though, if
they added new classes later, those would be sorted too low.

Also, note to self: slot_extension is probably only populated if there's
something in the slot. It was actually added with VERSION E REVISION 2
so the bool data structure's potentially been there the whole time. The
web site template language seems to check it before checking that the
slots are within slotNum, so that makes sense.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>